requirements.txt
# Python 3.13 compatible requirements - Updated June 2025
annotated-types>=0.7.0
anyio>=4.9.0
attrs>=23.2.0
blinker>=1.9.0
Brotli>=1.1.0
certifi>=2025.4.26
charset-normalizer>=3.4.2
click>=8.1.8
Deprecated>=1.2.18
distro>=1.9.0
flasgger>=0.9.7.1
Flask>=3.0.0
Flask-Assets>=2.1.0
Flask-Compress>=1.17
Flask-Limiter>=3.11.0
flask-talisman>=1.1.0
gunicorn>=23.0.0
h11>=0.16.0
httpcore>=1.0.9
httpx>=0.28.1
idna>=3.10
iniconfig>=2.1.0
itsdangerous>=2.2.0
Jinja2>=3.1.6
jiter>=0.10.0
jsonschema>=4.24.0
jsonschema-specifications>=2025.4.1
limits>=3.14.1
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
mdurl>=0.1.2
mistune>=3.1.3
numpy>=2.0.0
openai>=1.80.0
ordered-set>=4.1.0
packaging>=25.0
pandas>=2.2.0
pluggy>=1.6.0
pydantic>=2.10.0
pydantic_core>=2.27.0
Pygments>=2.19.1
pytest>=8.0.0
python-dateutil>=2.9.0
pytz>=2023.4
PyYAML>=6.0.2
RapidFuzz>=3.6.2
referencing>=0.30.2
requests>=2.32.2
rich>=13.9.4
rpds-py>=0.25.1
six>=1.17.0
sniffio>=1.3.1
tqdm>=4.67.1
typing-inspection>=0.4.1
typing_extensions>=4.9.0
tzdata>=2025.2
urllib3>=2.4.0
webassets>=2.0
Werkzeug>=3.1.3
wrapt>=1.17.2
zstandard>=0.23.0
# New AI Provider Dependencies
anthropic>=0.34.0
google-generativeai>=0.8.0
groq>=0.9.0
mistralai>=1.0.0
